+++ lib/napa.rb
@@ -42,11 +42,11 @@ require 'napa/deprecations'
 require 'napa/deploy'
 require 'napa/gem_dependency'
 
-# load rake tasks if Rake installed
 if defined?(Rake)
   load 'tasks/deploy.rake'
   load 'tasks/routes.rake'
-  load 'tasks/db.rake'
+
+  require 'tasks/database'
 end
 
 module Napa